Add files via upload
please download and install the following APK if you wish to participate in the evaluation of the mobile learning application. It will be considered a third party application and will, therefore, require you to change the installation permission on your device. (This can be set back to default once installation is complete) . The following links detail how to download and install the APK and also how to change device permission.

http://www.android.pk/blog/tutorials/how-to-enable-third-party-apps-installation-on-android-phones/
http://www.androidauthority.com/how-to-install-non-market-third-party-apps-on-your-phone-31494/